Add ESLint, refactor CacheableImage module
