+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "airbnb",
+  "rules": {
+    "react/jsx-filename-extension": ["error", { "extensions": [".js", ".jsx"] }],
+  }
+}
